Re #20 encode and decode number of spectra in messages

+++ event_data/include/DetectorSpectrumMapData.h
@@ -11,18 +11,18 @@ public:
   flatbuffers::unique_ptr_t getBufferPointer(std::string &buffer);
   void decodeMessage(const uint8_t *buf);
 
-  size_t getNumberOfEntries() { return m_numberOfEntries; }
+  int32_t getNumberOfEntries() { return m_numberOfEntries; }
   std::vector<int32_t> getDetectors() { return m_detectors; }
   std::vector<int32_t> getSpectra() { return m_spectra; }
   size_t getBufferSize() { return m_bufferSize; }
 
-  void setNumberOfEntries(size_t numberOfEntries) {
+  void setNumberOfEntries(int32_t numberOfEntries) {
     m_numberOfEntries = numberOfEntries;
   }
 
 private:
   void readFile(const std::string &filename);
-  size_t m_numberOfEntries = 0;
+  int32_t m_numberOfEntries = 0;
   size_t m_bufferSize = 0;
   std::vector<int32_t> m_detectors;
   std::vector<int32_t> m_spectra;
